# openapi.yaml
openapi: 3.0.0
info:
  title: StihiRus Reader API
  description: >
    Provides access to author data and poems from stihirus.ru with advanced caching (PostgreSQL),
    full-text search, filtering, and system monitoring.
  version: 1.6.1
servers:
  - url: /
    description: Current server
paths:
  /author/{identifier}:
    get:
      summary: Get Author Data, Poems, and Search
      description: >
        Fetches author profile and poems. Supports "Smart Sync" (incremental fetching),
        searching by text, filtering by year/rubric, and sorting.
        If `page` is provided, fetches specific page from source.
        If `page` is null/missing, syncs new content and returns full database result matching filters.
      parameters:
        - name: identifier
          in: path
          required: true
          description: Author identifier (numeric ID or string username).
          schema:
            type: string
            example: "oreh-orehov"
        
        # --- Standard Fetch Params ---
        - name: page
          in: query
          required: false
          description: >
            Specific page number to fetch from source (e.g., "1", "5").
            If omitted or null, "Smart Sync" mode is used (checks DB, fetches only new items).
          schema:
            type: string
            example: "1"
        - name: delay
          in: query
          required: false
          description: Delay in ms between requests during syncing. Defaults to 500ms.
          schema:
            type: integer
            default: 500

        # --- Search & Filter Params ---
        - name: q
          in: query
          required: false
          description: Text search query. Matches against fields specified in `searchFields`.
          schema:
            type: string
            example: "любовь"
        - name: searchFields
          in: query
          required: false
          description: >
            Comma-separated fields to search in. Default: "title,text".
            Allowed: "title", "text", "rubric".
          schema:
            type: string
            example: "title,text"
        - name: year
          in: query
          required: false
          description: Filter poems by publication year.
          schema:
            type: string
            example: "2024"
        - name: rubric
          in: query
          required: false
          description: Filter by exact rubric name.
          schema:
            type: string
            example: "Любовная лирика"
        - name: collection
          in: query
          required: false
          description: Filter by collection name (from metadata).
          schema:
            type: string
        
        # --- Sorting ---
        - name: sort
          in: query
          required: false
          description: >
            Sorting logic. Default: "date_desc".
            Options: 
            - "date_desc" (Newest first)
            - "date_asc" (Oldest first)
            - "popular" (or "rating_desc" - Most likes first)
            - "rating_asc" (Least likes first)
            - "title_asc" (A-Z)
            - "title_desc" (Z-A)
          schema:
            type: string
            enum: [date_desc, date_asc, popular, rating_desc, rating_asc, title_asc, title_desc]
            default: date_desc

      responses:
        '200':
          description: Successful response with author data and filtered poems.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StihirusSuccessResponse'
        '404':
          description: Author not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StihirusErrorResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StihirusErrorResponse'
  /author/{identifier}/filters:
    get:
      summary: Get Available Filters
      description: Fetches available rubrics and years for an author to build UI filters.
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StihirusFiltersSuccessResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StihirusErrorResponse'
  /poem/{id}:
    get:
      summary: Get Single Poem
      description: Fetches full details for a specific poem by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StihirusSinglePoemSuccessResponse'
        '404':
          description: Poem not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StihirusErrorResponse'
...
components:
  schemas:
    StihirusError:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required: [code, message]
    StihirusErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        error:
          $ref: '#/components/schemas/StihirusError'
      required: [status, error]
    StihirusAuthorStats:
      type: object
      properties:
        poems:
          type: integer
        reviewsSent:
          type: integer
        reviewsReceived:
          type: integer
        subscribers:
          type: integer
      required: [poems, reviewsSent, reviewsReceived]
    StihirusPoem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        text:
          type: string
        created:
          type: string
        rubric:
          type: object
          properties:
            name: { type: string }
            url: { type: string, nullable: true }
        collection:
          type: string
          nullable: true
        rating:
          type: integer
        commentsCount:
          type: integer
        imageUrl:
          type: string
          nullable: true
        hasCertificate:
          type: boolean
        gifts:
          type: array
          items:
            type: string
        uniquenessStatus:
          type: integer
          description: "-1: checking, 0: not unique, 1: unique"
        contest:
          type: object
          nullable: true
          properties:
            id: { type: integer }
            name: { type: string }
        holidaySection:
          type: object
          nullable: true
          properties:
            id: { type: integer }
            title: { type: string }
            url: { type: string }
      required: [id, title, text, created, rating]
    StihirusAuthorData:
      type: object
      properties:
        authorId: { type: integer }
        username: { type: string }
        profileUrl: { type: string }
        canonicalUsername: { type: string }
        description: { type: string }
        avatarUrl: { type: string, nullable: true }
        headerUrl: { type: string, nullable: true }
        status: { type: string }
        lastVisit: { type: string }
        isPremium: { type: boolean }
        stats:
          $ref: '#/components/schemas/StihirusAuthorStats'
        collections:
          type: array
          items:
            type: object # Simplified for brevity, detailed structure available if needed
        poems:
          type: array
          items:
            $ref: '#/components/schemas/StihirusPoem'
      required: [username, profileUrl, stats, poems]
    StihirusSuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: '#/components/schemas/StihirusAuthorData'
      required: [status, data]
    StihirusSinglePoemSuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: '#/components/schemas/StihirusPoem'
      required: [status, data]
    StihirusAuthorFiltersData:
      type: object
      properties:
        rubrics:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
              count: { type: integer }
        dates:
          type: array
          items:
            type: object
            properties:
              year: { type: integer }
              month: { type: integer }
              count: { type: integer }
      required: [rubrics, dates]
    StihirusFiltersSuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: '#/components/schemas/StihirusAuthorFiltersData'
      required: [status, data]
...